Built a new database abstaction.
I realized I'd never track down the problem without abstracting out the
database object from the tables. I did that, and implemented
environment fetching, and adjusted the tests to reflect those changes.
Additionally, added a lot of verbosity to the logging to make sure that
changes made to the db were persisting -- this should be toggle-able.